+++ Space Launches/Common/API/WebServiceAgent.swift
@@ -13,14 +13,15 @@ struct WebServiceAgent {
 
     private let networkManager = NetworkReachabilityManager()
 
-    func run<T: Decodable>(_ url: URLConvertible) -> AnyPublisher<[T], WebServiceAgentError> {
+    func run<T: Decodable>(_ url: URLConvertible) -> AnyPublisher<T, WebServiceAgentError> {
 
         if networkManager?.isReachable ?? false {
 
             return AF.request(url, method: .get)
                 .publishData()
+                .setFailureType(to: WebServiceAgentError.self)
                 .compactMap { $0.data }
-                .decode(type: [T].self, decoder: JSONDecoder())
+                .decode(type: T.self, decoder: JSONDecoder())
                 .receive(on: DispatchQueue.main)
                 .mapError { error in
                     switch error.asAFError {
@@ -30,6 +31,7 @@ struct WebServiceAgent {
                         return .invalidServerResponse
                     }
                 }
+                .timeout(.seconds(15), scheduler: DispatchQueue.main, customError: { .timeout })
                 .eraseToAnyPublisher()
         }