+++ Space Launches/Common/Protocols/ViewModelType.swift
@@ -1,0 +1,12 @@
+//
+//  ViewModelType.swift
+//  Space Launches
+//
+//  Created by Dominik Kohlman on 02/03/2021.
+//
+
+import Foundation
+
+protocol ViewModelType {
+    func send(event: UIFeedbacks.Event)
+}