+++ Space Launches/Common/ViewComponents/SpinnerView.swift
@@ -5,7 +5,6 @@
 //  Created by Dominik Kohlman on 21.02.2021.
 //
 
-import Foundation
 import SwiftUI
 
 struct SpinnerView: View {